Started modularizing
