+++ tools/data-conversion/src/results.txt
@@ -1,0 +1,7 @@
+
+| Command                     | Mean [s]      | Min [s] | Max [s] | Relative    |
+|-----------------------------|---------------|---------|---------|-------------|
+| Intermediate Representation | 3.607 ± 0.056 | 3.569   | 3.752   | 2.91 ± 0.29 |
+| Float Arithmetic            | 1.323 ± 0.050 | 1.286   | 1.460   | 1.07 ± 0.11 |
+| Bit Slicing                 | 1.239 ± 0.121 | 1.156   | 1.523   | 1.00        |
+
